config/addons/powersafe_tcp/requirements.txt
PyYAML
wakeonlan
requests
